+++ server/src/main/java/com/planmate/server/domain/Subject.java
@@ -21,7 +21,7 @@ import java.util.Calendar;
 @AllArgsConstructor
 public class Subject {
     @Id
-    @Column(name = "id",columnDefinition = "int")
+    @Column(name = "id",columnDefinition = "bigint")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @ApiModelProperty(name = "과목 태그 식별자")
     private Long id;
